requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
httptools==0.6.4
sqlalchemy==2.0.23

pymysql==1.1.0
cryptography
python-dotenv==1.0.1

# Auth / Sessions
bcrypt==4.0.1
passlib[bcrypt]
python-jose==3.3.0
itsdangerous==2.1.2

# Templates / Forms
jinja2==3.1.2
python-multipart==0.0.6

# Config / Scheduler
apscheduler==3.10.1

# Validation / Data
pydantic==2.7.1
pycountry==24.6.1
numpy==1.26.4
pandas==2.1.4

# Reports / Holidays
reportlab==4.2.2
holidays==0.55

# Required for FastAPI and middleware
starlette==0.32.0
